+++ packages/react-ui-kit/src/Icon/Brand/FacebookIcon.jsx
@@ -17,13 +17,13 @@
  *
  */
 
-import IconHOC from './IconHOC';
+import IconHOC from '../IconHOC';
 import React from 'react';
 
 const size = 16;
-const twitter = (
-  <path d="M13.73 3.687a3.079 3.079 0 0 0 1.353-1.71 6.27 6.27 0 0 1-1.944.75 3.14 3.14 0 0 0-2.257-.977 3.075 3.075 0 0 0-3.073 3.089c0 .244.017.488.07.715-2.553-.14-4.81-1.309-6.337-3.176-.26.454-.417.925-.417 1.5a3.08 3.08 0 0 0 1.372 2.566 3.118 3.118 0 0 1-1.39-.384c0 1.536 1.06 2.792 2.466 3.071a3.17 3.17 0 0 1-1.389.053c.382 1.221 1.528 2.129 2.865 2.146-1.042.838-2.379 1.466-3.82 1.466-.243 0-.486-.018-.729-.052C1.854 13.616 3.486 14 5.222 14c5.66 0 8.75-4.712 8.75-8.795v-.401A6.391 6.391 0 0 0 15.5 3.198a6.06 6.06 0 0 1-1.77.489z" />
+const facebook = (
+  <path d="M14.227 1H1.773A.773.773 0 0 0 1 1.773v12.454c0 .427.346.773.773.773h6.705V9.578H6.653V7.466h1.825V5.907c0-1.808 1.104-2.793 2.717-2.793a14.8 14.8 0 0 1 1.63.084v1.89h-1.118c-.877 0-1.047.417-1.047 1.029v1.349h2.092l-.272 2.112h-1.82V15h3.567a.773.773 0 0 0 .773-.773V1.773A.773.773 0 0 0 14.227 1" />
 );
-const TwitterIcon = IconHOC(twitter, size, size);
+const FacebookIcon = IconHOC(facebook, size, size);
 
-export {TwitterIcon};
+export {FacebookIcon};
