+++ packages/react-ui-kit/src/Icon/Brand/LinkedInIcon.jsx
@@ -17,13 +17,13 @@
  *
  */
 
-import IconHOC from './IconHOC';
+import IconHOC from '../IconHOC';
 import React from 'react';
 
 const size = 16;
-const facebook = (
-  <path d="M14.227 1H1.773A.773.773 0 0 0 1 1.773v12.454c0 .427.346.773.773.773h6.705V9.578H6.653V7.466h1.825V5.907c0-1.808 1.104-2.793 2.717-2.793a14.8 14.8 0 0 1 1.63.084v1.89h-1.118c-.877 0-1.047.417-1.047 1.029v1.349h2.092l-.272 2.112h-1.82V15h3.567a.773.773 0 0 0 .773-.773V1.773A.773.773 0 0 0 14.227 1" />
+const linkedin = (
+  <path d="M.333 5.333v10.334h3.334V5.333H.333zM9 9.667c0-1.334.852-2 1.852-2s1.815.666 1.815 2.333v5.667H16v-6C16 6.667 14.333 5 12 5c-1.333 0-2.333.76-3 1.76l-.111-1.427H5.61c0 .352.056 2.334.056 2.334v8H9v-6zm-7.019-6C3.222 3.667 4 2.852 4 1.833 3.981.796 3.222 0 2.019 0 .796 0 0 .796 0 1.833c0 1.019.778 1.834 1.963 1.834h.018z" />
 );
-const FacebookIcon = IconHOC(facebook, size, size);
+const LinkedInIcon = IconHOC(linkedin, size, size);
 
-export {FacebookIcon};
+export {LinkedInIcon};
