+++ packages/react-ui-kit/src/Icon/Brand/TwitterIcon.jsx
@@ -17,13 +17,13 @@
  *
  */
 
-import IconHOC from './IconHOC';
+import IconHOC from '../IconHOC';
 import React from 'react';
 
 const size = 16;
-const linkedin = (
-  <path d="M.333 5.333v10.334h3.334V5.333H.333zM9 9.667c0-1.334.852-2 1.852-2s1.815.666 1.815 2.333v5.667H16v-6C16 6.667 14.333 5 12 5c-1.333 0-2.333.76-3 1.76l-.111-1.427H5.61c0 .352.056 2.334.056 2.334v8H9v-6zm-7.019-6C3.222 3.667 4 2.852 4 1.833 3.981.796 3.222 0 2.019 0 .796 0 0 .796 0 1.833c0 1.019.778 1.834 1.963 1.834h.018z" />
+const twitter = (
+  <path d="M13.73 3.687a3.079 3.079 0 0 0 1.353-1.71 6.27 6.27 0 0 1-1.944.75 3.14 3.14 0 0 0-2.257-.977 3.075 3.075 0 0 0-3.073 3.089c0 .244.017.488.07.715-2.553-.14-4.81-1.309-6.337-3.176-.26.454-.417.925-.417 1.5a3.08 3.08 0 0 0 1.372 2.566 3.118 3.118 0 0 1-1.39-.384c0 1.536 1.06 2.792 2.466 3.071a3.17 3.17 0 0 1-1.389.053c.382 1.221 1.528 2.129 2.865 2.146-1.042.838-2.379 1.466-3.82 1.466-.243 0-.486-.018-.729-.052C1.854 13.616 3.486 14 5.222 14c5.66 0 8.75-4.712 8.75-8.795v-.401A6.391 6.391 0 0 0 15.5 3.198a6.06 6.06 0 0 1-1.77.489z" />
 );
-const LinkedInIcon = IconHOC(linkedin, size, size);
+const TwitterIcon = IconHOC(twitter, size, size);
 
-export {LinkedInIcon};
+export {TwitterIcon};
